Migrate benchmarks to WASI-SDK compilation and consolidate build, test, and documentation commands under Makefile
