+++ .github/workflows/go.yml
@@ -40,11 +40,6 @@ jobs:
         tar -xzf wabt-${WABT_VERSION}-linux-x64.tar.gz
         echo "$(pwd)/wabt-${WABT_VERSION}/bin" >> $GITHUB_PATH
     - name: Build
-      run: |
-        go build ./...
-        GOOS=darwin go build -o epsilon-darwin ./cmd/epsilon
-        GOOS=windows go build -o epsilon.exe ./cmd/epsilon
+      run: make build-all
     - name: Test
-      run: go test ./...
-    - name: Run WASI Tests
-      run: uv run --with-requirements requirements.txt wasip1/wasi_testsuite.py
+      run: make test-all
