+++ .gitignore
@@ -26,3 +26,9 @@ __pycache__/
 .venv/
 *.pyc
 .antigravitycli
+
+# Local toolchain installs (wasi-sdk, etc.)
+.toolchain/
+
+# Benchmark .wasm artifacts — built on demand via `make build-wasm`
+internal/benchmarks/wasm/
