+++ wasip1/README.md
@@ -73,7 +73,13 @@ This implementation is tested against the official
 
 ### Running WASI Spec Tests
 
-From the repository root:
+From the repository root, you can run the tests using `make`:
+
+```bash
+make test-wasi
+```
+
+Or run the underlying command directly:
 
 ```bash
 uv run --with-requirements requirements.txt wasip1/wasi_testsuite.py
